% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cover_page.R
\name{add_cover_page}
\alias{add_cover_page}
\title{Cover page}
\usage{
add_cover_page(..., header_color = "#f44242", footer_color = "#3d3c3c",
  creaton_time = Sys.Date(), logo = NA, logo_size = NULL,
  theme = c("basic", "flashy"))
}
\arguments{
\item{...}{Other arguments passed on to methods.
It is only for technical usage.}

\item{header_color}{Character string which specifies the color of the header.
Only HEX color codes are accepted (e.g. #ffffff). Default value is "#f44242".}

\item{footer_color}{Character string which specifies the color of the footer.
Only HEX color codes are accepted (e.g. #ffffff). Default value is "#3d3c3c".}

\item{creaton_time}{Character string value which is implemented for
displaying creation time of the report but it can be used for other purpose.
Default value is Sys.Date().}

\item{logo}{Name of the logo file which has been already loaded into R.
I recommend to use image_read function from package magick
to load images into R. Default value is NA which means that nothing
will appear in place of the logo.}

\item{logo_size}{Numeric value which specifies the size of the logo.
Default value is NULL which means that the original size of the logo
file will be used.}

\item{theme}{Character string which specifies the theme of the report.
Possible values are "basic" and "flashy". Try both
themes and check the differences.}
}
\description{
This function generates the cover page of the report. It is fully optional
to use it so just leave this function out of the code if you do not need
cover page.
}
\examples{
# minimal example
# it generates PDF file with only a cover page.
start_report() \%>\%
  add_cover_page() \%>\%
  end_report()

# change the theme of the report from default.
start_report() \%>\%
  add_cover_page(theme = "flashy") \%>\%
  end_report()

# add a logo to a report
# 1. load the logo file.
library(magick)
sample_logo <- image_read("http://myhappydata.com/img/1.png")
# 2. add the loaded file to a report.
start_report() \%>\%
  add_cover_page(logo = sample_logo) \%>\%
  end_report()

# change the default colors to custom ones.
start_report() \%>\%
  add_cover_page(
    header_color = "#6DCFF6",
    footer_color = "#D8D3CB"
  ) \%>\%
  end_report()

}
