% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_report.R
\name{end_report}
\alias{end_report}
\title{End report}
\usage{
end_report(...)
}
\arguments{
\item{...}{Other arguments passed on to methods.
It is only for technical usage.}
}
\description{
This function is not requires any values to specify.
It is only implemented for generating the final output file and cleaning up
after the report making process ends. Important to note that this only
works if at least the start_report() function has been executed already.
}
\examples{
# minimal example
# it generates a PDF file which does not contains anything
start_report() \%>\%
  end_report()
}
