% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_report.R
\name{start_report}
\alias{start_report}
\title{Start report}
\usage{
start_report(filename = "report_output", title = "Coolest Report Ever",
  subtitle = "Easy & flexible way to create good-looking reports.",
  size = "a4")
}
\arguments{
\item{filename}{Character string giving the name of the output PDF file.}

\item{title}{Title of the report. Character string which appears both on the
cover page and the report page. Default title is "Coolest Report Ever".}

\item{subtitle}{Subitle of the report. Character string which appears both
on the cover page and the report page. Default
subtitle is "Easy & flexible way to create good-looking reports.".}

\item{size}{Character string which specifies the paper size of the output
PDF file. Currently implemented values
are "a4" and "a4r" (rotated landscape). Default paper size is "a4".}
}
\description{
This is the first mandatory function to generate the report.
It defines some basic parameters.
}
\examples{
# minimal example
# it generates a PDF file which does not contains anything
start_report() \%>\%
  end_report()

# custom filename, title and subtitle on the report page
start_report(
    filename = "mycustom_output",
    title = "My Custom Title",
    subtitle = "My Custom Subtitle"
) \%>\%
  add_new_page() \%>\%
  end_report()

# rotated landscape (a4r)
start_report(size = "a4r") \%>\%
  add_new_page() \%>\%
  end_report()
}
